+++ examples/data_pipeline/pipelines.py
@@ -6,6 +6,9 @@ from transforms import (
     PyTextVocabTransform,
     PyTextScriptVocabTransform,
     iterate_batch,
+    tokenizer_func,
+    totensor,
+    vocab_func,
 )
 from torchtext.experimental.transforms import (
     basic_english_normalize,
@@ -14,9 +17,6 @@ from torchtext.experimental.transforms import (
 from torchtext.data.utils import get_tokenizer
 from torchtext.experimental.functional import (
     sequential_transforms,
-    tokenizer_func,
-    totensor,
-    vocab_func,
 )
 from torchtext.experimental.vectors import FastText as FastTextExperimental
 from torchtext.experimental.vocab import vocab_from_file
@@ -50,7 +50,7 @@ def build_legacy_torchtext_vocab_pipeline(vocab_file):
         for line in f:
             for token in line:
                 yield token
-                
+
     vocab = build_vocab_from_iterator(token_iterator(vocab_file))
     pipeline = sequential_transforms(tokenizer_func(tokenizer), vocab_func(vocab))
     return iterate_batch(pipeline), None, None
@@ -106,7 +106,7 @@ def build_legacy_pytext_script_vocab_pipeline(vocab_file):
 
     tokenizer = basic_english_normalize()
     f = open(vocab_file, 'r')
-    
+
     vocab_counter = Counter([token for line in f for token in line.rstrip()])
     sorted_by_freq_tuples = sorted(vocab_counter.items(), key=lambda x: x[1], reverse=True)
     vocab_list = [pair[0] for pair in sorted_by_freq_tuples]
@@ -141,7 +141,7 @@ def build_experimental_pytext_script_vocab_pipeline(vocab_file):
 
 
 def build_legacy_fasttext_vector_pipeline():
-    from torchtext.experimental.functional import vector_func
+    from transforms import vector_func
     tokenizer = get_tokenizer("basic_english")
     vector = FastText()
